requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
scikit-learn>=1.0.0
xgboost>=1.5.0
lightgbm>=3.3.0
torch>=1.10.0
pyyaml>=5.4.0
joblib>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0

